requirements.txt
pydirectinput==1.0.4
PyYAML==6.0.1
keyboard==0.13.5
mido==1.3.2
PySide6==6.7.0 